% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read_years.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Add and select columns on datasets}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{integer or integer vector specifying year of the dataset
ready to be read in.}
}
\value{
List of tibbles containing columns of year and month
}
\description{
Add new column of year data, select columns related to month
and year data in CSV files containing American public yearly data
regarding fatal injuries suffered in motor vehicle traffic crashes,
then convert the default dataframe format to tibble.
}
\details{
\itemize{
 \item{}{Valid years are exclusively 2013, 2014 and 2015, an error
 message will be post and program will be stopped if invalid years
 are inputted}
 \item{}{User is required to install prerequisted package \code{\link{dplyr}},
 an error message will be post and program will be stopped}
 }
}
\examples{
\dontrun{
years <- c(2013,2014,2015)
fars_read_years(years)}

}
