% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize the counts of occurence}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{integer or integer vectors specifying year
of the dataset ready to be read in}
}
\value{
Tibble summarizing the counts of occurence by
month for the year
}
\description{
Create dataframe to summarize the counts of occurence of
fatal injuries suffered in motor vehicle traffic crashes.
}
\details{
\itemize{
 \item{}{Valid years are exclusively 2013, 2014 and 2015, an error
 message will be post and program will be stopped if invalid years
 are inputted}
 \item{}{User is required to install prerequisted package \code{\link{dplyr}}
 and \code{\link{tidyr}}, an error message will be post and program will be stopped}
 }
}
\examples{
\dontrun{years <- c(2013,2014,2015)
fars_summarize_years(years)}
}
