% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YGWLGiaoGiao.R
\docType{package}
\name{YGWLGiaoGiao}
\alias{YGWLGiaoGiao}
\alias{YGWLGiaoGiao-package}
\title{YGWLGiaoGiao package}
\description{
Contians five functions for for analyzing data from the US National Highway Traffic Safety Administration's Fatality Analysis
}
\author{
Xinwei Long \email{jasonxinweilong@gmail.com}
}
